+++ cmd/ttn-lw-cli/commands/applications_claim.go
@@ -38,15 +38,30 @@ key is provided, a new API key will be created.`,
 				return errNoApplicationID
 			}
 
+			expiryDate, err := getAPIKeyExpiry(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
 			key, _ := cmd.Flags().GetString("api-key")
 			if key == "" {
+				is, err := api.Dial(ctx, config.IdentityServerGRPCAddress)
+				if err != nil {
+					return err
+				}
 				logger.Info("Creating API key")
-				apiKey, err := createApplicationAPIKey(ctx, *appID, "Device Claiming",
-					ttnpb.RIGHT_APPLICATION_DEVICES_READ,
-					ttnpb.RIGHT_APPLICATION_DEVICES_READ_KEYS,
-					ttnpb.RIGHT_APPLICATION_DEVICES_WRITE,
-					ttnpb.RIGHT_APPLICATION_DEVICES_WRITE_KEYS,
-				)
+				apiKey, err := ttnpb.NewApplicationAccessClient(is).CreateAPIKey(ctx, &ttnpb.CreateApplicationAPIKeyRequest{
+					ApplicationIdentifiers: *appID,
+					Name:                   "Device Claiming",
+					Rights: []ttnpb.Right{
+						ttnpb.RIGHT_APPLICATION_DEVICES_READ,
+						ttnpb.RIGHT_APPLICATION_DEVICES_READ_KEYS,
+						ttnpb.RIGHT_APPLICATION_DEVICES_WRITE,
+						ttnpb.RIGHT_APPLICATION_DEVICES_WRITE_KEYS,
+					},
+					ExpiresAt: expiryDate,
+				})
+
 				if err != nil {
 					return err
 				}
@@ -88,6 +103,7 @@ key is provided, a new API key will be created.`,
 
 func init() {
 	applicationClaimAuthorize.Flags().String("api-key", "", "")
+	applicationClaimAuthorize.Flags().String("api-key-expiry", "", "API key expiry date (YYYY-MM-DD:HH:mm) - only applicable when creating API Key")
 	applicationClaim.AddCommand(applicationClaimAuthorize)
 	applicationClaim.AddCommand(applicationClaimUnauthorize)
 	applicationClaim.PersistentFlags().AddFlagSet(applicationIDFlags())