+++ pkg/identityserver/store/api_key_store.go
@@ -16,10 +16,14 @@ package store
 
 import (
 	"context"
+	"fmt"
 	"runtime/trace"
+	"strings"
 
+	"github.com/gogo/protobuf/types"
 	"github.com/jinzhu/gorm"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
+	"go.thethings.network/lorawan-stack/v3/pkg/rpcmiddleware/warning"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 )
 
@@ -30,6 +34,30 @@ func GetAPIKeyStore(db *gorm.DB) APIKeyStore {
 
 type apiKeyStore struct {
 	*store
+}
+
+func selectAPIKeyFields(ctx context.Context, query *gorm.DB, fieldMask *types.FieldMask) *gorm.DB {
+	var apiKeyColumns []string
+	var notFoundPaths []string
+
+	for _, path := range ttnpb.TopLevelFields(fieldMask.Paths) {
+		switch path {
+		case "updated_at", "entity_id", "entity_type", "id":
+			// always selected
+		case "expires_at":
+			apiKeyColumns = append(apiKeyColumns, "expires_at")
+		case "rights":
+			apiKeyColumns = append(apiKeyColumns, "rights")
+		case "name":
+			apiKeyColumns = append(apiKeyColumns, "name")
+		default:
+			notFoundPaths = append(notFoundPaths, path)
+		}
+	}
+	if len(notFoundPaths) > 0 {
+		warning.Add(ctx, fmt.Sprintf("unsupported field mask paths: %s", strings.Join(notFoundPaths, ", ")))
+	}
+	return query.Select(cleanFields(append(append(modelColumns, "updated_at"), apiKeyColumns...)...))
 }
 
 func (s *apiKeyStore) CreateAPIKey(ctx context.Context, entityID *ttnpb.EntityIdentifiers, key *ttnpb.APIKey) (*ttnpb.APIKey, error) {
@@ -45,6 +73,7 @@ func (s *apiKeyStore) CreateAPIKey(ctx context.Context, entityID *ttnpb.EntityId
 		Name:       key.Name,
 		EntityID:   entity.PrimaryKey(),
 		EntityType: entityTypeForID(entityID),
+		ExpiresAt:  key.ExpiresAt,
 	}
 	if err = s.createEntity(ctx, model); err != nil {
 		return nil, err
@@ -106,7 +135,7 @@ func (s *apiKeyStore) GetAPIKey(ctx context.Context, id string) (*ttnpb.EntityId
 	return ids, keyModel.toPB(), nil
 }
 
-func (s *apiKeyStore) UpdateAPIKey(ctx context.Context, entityID *ttnpb.EntityIdentifiers, key *ttnpb.APIKey) (*ttnpb.APIKey, error) {
+func (s *apiKeyStore) UpdateAPIKey(ctx context.Context, entityID *ttnpb.EntityIdentifiers, key *ttnpb.APIKey, fieldMask *types.FieldMask) (*ttnpb.APIKey, error) {
 	defer trace.StartRegion(ctx, "update api key").End()
 	entity, err := s.findEntity(ctx, entityID, "id")
 	if err != nil {
@@ -125,12 +154,25 @@ func (s *apiKeyStore) UpdateAPIKey(ctx context.Context, entityID *ttnpb.EntityId
 		}
 		return nil, err
 	}
-	if len(key.Rights) == 0 {
+	// Support for previous versions of The Things Stack.
+	if len(fieldMask.Paths) == 0 {
+		fieldMask.Paths = []string{"rights", "name"}
+	}
+	// If empty rights are passed and rights are in the fieldmask, delete the key.
+	if len(key.Rights) == 0 && ttnpb.HasAnyField(fieldMask.Paths, "rights") {
 		return nil, query.Delete(&keyModel).Error
 	}
-	keyModel.Name = key.Name
-	keyModel.Rights = Rights{Rights: key.Rights}
-	if err = query.Select("name", "rights", "updated_at").Save(&keyModel).Error; err != nil {
+	query = selectAPIKeyFields(ctx, query, fieldMask)
+	if ttnpb.HasAnyField(fieldMask.Paths, "rights") {
+		keyModel.Rights = Rights{Rights: key.Rights}
+	}
+	if ttnpb.HasAnyField(fieldMask.Paths, "expires_at") {
+		keyModel.ExpiresAt = key.ExpiresAt
+	}
+	if ttnpb.HasAnyField(fieldMask.Paths, "name") {
+		keyModel.Name = key.Name
+	}
+	if err = query.Save(&keyModel).Error; err != nil {
 		return nil, err
 	}
 	return keyModel.toPB(), nil