+++ pkg/ttnpb/application.pb.paths.fm.go
@@ -131,18 +131,21 @@ var GetApplicationAPIKeyRequestFieldPathsTopLevel = []string{
 var CreateApplicationAPIKeyRequestFieldPathsNested = []string{
 	"application_ids",
 	"application_ids.application_id",
+	"expires_at",
 	"name",
 	"rights",
 }
 
 var CreateApplicationAPIKeyRequestFieldPathsTopLevel = []string{
 	"application_ids",
+	"expires_at",
 	"name",
 	"rights",
 }
 var UpdateApplicationAPIKeyRequestFieldPathsNested = []string{
 	"api_key",
 	"api_key.created_at",
+	"api_key.expires_at",
 	"api_key.id",
 	"api_key.key",
 	"api_key.name",
@@ -150,11 +153,13 @@ var UpdateApplicationAPIKeyRequestFieldPathsNested = []string{
 	"api_key.updated_at",
 	"application_ids",
 	"application_ids.application_id",
+	"field_mask",
 }
 
 var UpdateApplicationAPIKeyRequestFieldPathsTopLevel = []string{
 	"api_key",
 	"application_ids",
+	"field_mask",
 }
 var ListApplicationCollaboratorsRequestFieldPathsNested = []string{
 	"application_ids",