+++ pkg/ttnpb/field_mask_validation.go
@@ -408,6 +408,9 @@ var RPCFieldMaskPaths = map[string]RPCFieldMaskPathValue{
 	"/ttn.lorawan.v3.As/GetLink": {All: asApplicationLinkFieldPaths, Allowed: asApplicationLinkFieldPaths},
 	"/ttn.lorawan.v3.As/SetLink": {All: asApplicationLinkFieldPaths, Allowed: asApplicationLinkFieldPaths, Set: true},
 
+	// Application API Keys:
+	"/ttn.lorawan.v3.ApplicationAccess/UpdateAPIKey": {All: APIKeyFieldPathsNested, Allowed: APIKeyFieldPathsNested, Set: true},
+
 	// Clients:
 	"/ttn.lorawan.v3.ClientRegistry/Get":                 {All: ClientFieldPathsNested, Allowed: omitFields(ClientFieldPathsNested, "secret")},
 	"/ttn.lorawan.v3.ClientRegistry/List":                {All: ClientFieldPathsNested, Allowed: omitFields(ClientFieldPathsNested, "secret")},
@@ -890,17 +893,26 @@ var RPCFieldMaskPaths = map[string]RPCFieldMaskPathValue{
 		Set: true,
 	},
 
+	// Gateway API Keys:
+	"/ttn.lorawan.v3.GatewayAccess/UpdateAPIKey": {All: APIKeyFieldPathsNested, Allowed: APIKeyFieldPathsNested, Set: true},
+
 	// Organizations:
 	"/ttn.lorawan.v3.OrganizationRegistry/Get":                 {All: OrganizationFieldPathsNested, Allowed: OrganizationFieldPathsNested},
 	"/ttn.lorawan.v3.OrganizationRegistry/List":                {All: OrganizationFieldPathsNested, Allowed: OrganizationFieldPathsNested},
 	"/ttn.lorawan.v3.OrganizationRegistry/Update":              {All: OrganizationFieldPathsNested, Allowed: OrganizationFieldPathsNested, Set: true},
 	"/ttn.lorawan.v3.EntityRegistrySearch/SearchOrganizations": {All: OrganizationFieldPathsNested, Allowed: OrganizationFieldPathsNested},
 
+	// Organization API Keys:
+	"/ttn.lorawan.v3.OrganizationAccess/UpdateAPIKey": {All: APIKeyFieldPathsNested, Allowed: APIKeyFieldPathsNested, Set: true},
+
 	// Users:
 	"/ttn.lorawan.v3.UserRegistry/Get":                 {All: UserFieldPathsNested, Allowed: omitFields(UserFieldPathsNested, "password", "temporary_password")},
 	"/ttn.lorawan.v3.UserRegistry/List":                {All: UserFieldPathsNested, Allowed: omitFields(UserFieldPathsNested, "password", "temporary_password")},
 	"/ttn.lorawan.v3.UserRegistry/Update":              {All: UserFieldPathsNested, Allowed: omitFields(UserFieldPathsNested, "password", "password_updated_at"), Set: true},
 	"/ttn.lorawan.v3.EntityRegistrySearch/SearchUsers": {All: UserFieldPathsNested, Allowed: omitFields(UserFieldPathsNested, "password", "temporary_password")},
+
+	// User API Keys:
+	"/ttn.lorawan.v3.UserAccess/UpdateAPIKey": {All: APIKeyFieldPathsNested, Allowed: APIKeyFieldPathsNested, Set: true},
 
 	// Storage Integration:
 	"/ttn.lorawan.v3.ApplicationUpStorage/GetStoredApplicationUp": {All: ApplicationUpFieldPathsNested, Allowed: applicationUpFieldMaskPaths()},