+++ pkg/ttnpb/organization.pb.validate.go
@@ -1061,6 +1061,21 @@ func (m *CreateOrganizationAPIKeyRequest) ValidateFields(paths ...string) error 
 
 			}
 
+		case "expires_at":
+
+			if ts := m.GetExpiresAt(); ts != nil {
+
+				now := time.Now()
+
+				if ts.Sub(now) <= 0 {
+					return CreateOrganizationAPIKeyRequestValidationError{
+						field:  "expires_at",
+						reason: "value must be greater than now",
+					}
+				}
+
+			}
+
 		default:
 			return CreateOrganizationAPIKeyRequestValidationError{
 				field:  name,
@@ -1161,6 +1176,18 @@ func (m *UpdateOrganizationAPIKeyRequest) ValidateFields(paths ...string) error 
 				if err := v.ValidateFields(subs...); err != nil {
 					return UpdateOrganizationAPIKeyRequestValidationError{
 						field:  "api_key",
+						reason: "embedded message failed validation",
+						cause:  err,
+					}
+				}
+			}
+
+		case "field_mask":
+
+			if v, ok := interface{}(&m.FieldMask).(interface{ ValidateFields(...string) error }); ok {
+				if err := v.ValidateFields(subs...); err != nil {
+					return UpdateOrganizationAPIKeyRequestValidationError{
+						field:  "field_mask",
 						reason: "embedded message failed validation",
 						cause:  err,
 					}